moved tweetUpdates to services folder
